requirements.txt
discord.py==2.4.0
python-dotenv==1.0.1
uvicorn==0.30.6
fastapi==0.112.1
